requirements.txt
streamlit
requests
python-dotenv
pyyaml
watchdog
pytest
flake8
fastapi
uvicorn
